Add unitest for intokens
