+++ paddlenlp/datasets/__init__.py
@@ -27,6 +27,7 @@ from .drcd_cn import *
 from .dureader_robust import *
 from .glue import *
 from .imdb import *
+from .intokens_dataset import *
 from .lcqmc import *
 from .msra_ner import *
 from .nlpcc13_evsam05_hit import *
